integration-tests/pyproject.toml
[project]
name = "tileserver-integration-tests"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pillow>=10.0",
    "pytest-timeout>=2.3",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-ra"
timeout = 120
asyncio_mode = "auto"
markers = [
    "correctness: functional correctness tests",
    "stability: long-running stability tests",
    "performance: throughput and latency benchmarks",
]
